requirements.txt
argparse
easydict
h5py
matplotlib
numpy
opencv-python
pyyaml
scipy
tensorboardX
timm
tqdm
termcolor
pandas
ninja
ftfy
regex
einops
scikit-learn
torch-scatter==2.0.9
